feat: update browser router
